requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.1
Flask==3.1.2
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.33.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langgraph==0.4.0
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.1.8
langgraph-sdk==0.2.6
langsmith==0.1.147
MarkupSafe==3.0.2
multidict==6.6.4
numpy==1.26.4
orjson==3.11.3
ormsgpack==1.10.0
packaging==24.2
Pillow==10.0.1
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.0
pydantic_core==2.14.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tenacity==8.5.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
celery==5.4.0
redis==5.0.8
python-crontab==3.2.0